배포

+++ src/main/java/com/b6/mypaldotrip/domain/city/service/CityService.java
@@ -14,9 +14,11 @@ import com.b6.mypaldotrip.domain.city.s3.repository.CityFileRepository;
 import com.b6.mypaldotrip.domain.city.store.entity.CityEntity;
 import com.b6.mypaldotrip.domain.city.store.entity.CitySort;
 import com.b6.mypaldotrip.domain.city.store.repository.CityRepository;
-import com.b6.mypaldotrip.domain.user.store.entity.UserEntity;
+import com.b6.mypaldotrip.domain.trip.exception.TripErrorCode;
+import com.b6.mypaldotrip.domain.user.store.entity.UserRole;
 import com.b6.mypaldotrip.global.common.S3Provider;
 import com.b6.mypaldotrip.global.exception.GlobalException;
+import com.b6.mypaldotrip.global.security.UserDetailsImpl;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
 import java.util.List;
@@ -33,15 +35,19 @@ public class CityService {
     private final CityFileRepository cityFileRepository;
     private final S3Provider s3Provider;
 
-    public CityCreateRes createCity(String req, UserEntity user, MultipartFile multipartFile)
+    public CityCreateRes createCity(
+            String req, UserDetailsImpl userDetails, MultipartFile multipartFile)
             throws IOException {
         ObjectMapper objectMapper = new ObjectMapper();
         CityCreateReq cityCreateReq = objectMapper.readValue(req, CityCreateReq.class);
         // 유저의 권한이 관리자인지 확인(securityconfig에서 관리자만 할수 있게 하는지)
-
+        checkAuthorization(userDetails);
         // 같은 시 명이 중복되는지 확인
         cityDuplicationCheck(cityCreateReq.cityName());
-
+        // MultipartFile이 null이 아니고 이미지가 아닌 경우 예외 처리
+        if (multipartFile != null && !isImage(multipartFile)) {
+            throw new GlobalException(CityErrorCode.WRONG_FILE_EXTENSION);
+        }
         CityEntity cityEntity =
                 CityEntity.builder()
                         .provinceName(cityCreateReq.provinceName())
@@ -49,11 +55,13 @@ public class CityService {
                         .cityInfo(cityCreateReq.cityInfo())
                         .build();
 
+        // MultipartFile이 null이 아니면 이미지 업로드 및 저장
+        if (multipartFile != null) {
+            String fileURL = s3Provider.saveFile(multipartFile, "city");
+            cityFileRepository.save(
+                    CityFileEntity.builder().fileUrl(fileURL).cityEntity(cityEntity).build());
+        }
         cityRepository.save(cityEntity);
-
-        String fileURL = s3Provider.saveFile(multipartFile, "city");
-        cityFileRepository.save(
-                CityFileEntity.builder().fileUrl(fileURL).cityEntity(cityEntity).build());
 
         return CityCreateRes.builder()
                 .provinceName(cityEntity.getProvinceName())
@@ -63,7 +71,8 @@ public class CityService {
     }
 
     @Transactional
-    public CityUpdateRes updateCity(Long cityId, CityUpdateReq req) {
+    public CityUpdateRes updateCity(Long cityId, CityUpdateReq req, UserDetailsImpl userDetails) {
+        checkAuthorization(userDetails);
         // 수정하려는 시가 존재하는지 확인
         CityEntity cityEntity = findCity(cityId);
         cityEntity.update(req.provinceName(), req.cityName(), req.cityInfo());
@@ -75,7 +84,8 @@ public class CityService {
                 .build();
     }
 
-    public CityDeleteRes deleteCity(Long cityId) {
+    public CityDeleteRes deleteCity(Long cityId, UserDetailsImpl userDetails) {
+        checkAuthorization(userDetails);
         CityEntity cityEntity = findCity(cityId);
         cityRepository.delete(cityEntity);
         CityDeleteRes res =
@@ -140,5 +150,28 @@ public class CityService {
         if (cityRepository.findByCityName(cityName).isPresent()) {
             throw new GlobalException(CityErrorCode.ALREADY_CITY_EXIST);
         }
+    }
+
+    private static void checkAuthorization(UserDetailsImpl userDetails) {
+        if (userDetails.getUserEntity().getUserRole() == UserRole.ROLE_USER) {
+            throw new GlobalException(TripErrorCode.UNAUTHORIZED_ROLE_ERROR);
+        }
+    }
+
+    public boolean isImage(MultipartFile file) {
+
+        String contentType = file.getContentType();
+
+        if (contentType != null && contentType.startsWith("image")) {
+            String fileName = file.getOriginalFilename();
+            if (fileName != null) {
+                String lowerCaseFileName = fileName.toLowerCase();
+                if (lowerCaseFileName.endsWith(".gif")) {
+                    return false; // 잘못된 확장자인 경우 false 반환
+                }
+                return true;
+            }
+        }
+        return false;
     }
 }
